Divide routes into list, todos, users
